requirements.txt
# Development Dependencies
Sphinx==3.4.3
sphinx-rtd-theme==0.5.1
yapf==0.30.0
Pillow==8.1.0
requests==2.24.0

# Production Dependencies
numpy==1.20.1
sense-hat==2.2.0
picamera==1.13